Renamed draw.js to drawEdge.js
